+++ flutter_inappwebview_platform_interface/lib/src/types/main.dart
@@ -246,3 +246,7 @@ export 'environment_release_channels.dart' show EnvironmentReleaseChannels;
 export 'environment_scrollbar_style.dart' show EnvironmentScrollbarStyle;
 export 'browser_process_exit_kind.dart' show BrowserProcessExitKind;
 export 'browser_process_exited_detail.dart' show BrowserProcessExitedDetail;
+export 'browser_process_kind.dart' show BrowserProcessKind;
+export 'browser_process_info.dart' show BrowserProcessInfo;
+export 'browser_process_infos_changed_detail.dart'
+    show BrowserProcessInfosChangedDetail;