+++ flutter_inappwebview_windows/lib/src/webview_environment/webview_environment.dart
@@ -73,6 +73,12 @@ class WindowsWebViewEnvironment extends PlatformWebViewEnvironment
           onBrowserProcessExited?.call(detail);
         }
         break;
+        case 'onProcessInfosChanged':
+        if (onProcessInfosChanged != null) {
+          Map<String, dynamic> arguments = call.arguments.cast<String, dynamic>();
+          final detail = BrowserProcessInfosChangedDetail.fromMap(arguments)!;
+          onProcessInfosChanged?.call(detail);
+        }
       default:
         throw UnimplementedError("Unimplemented ${call.method} method");
     }
@@ -84,6 +90,21 @@ class WindowsWebViewEnvironment extends PlatformWebViewEnvironment
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('interface', () => interface.toNativeValue());
     return await channel?.invokeMethod<bool>('isInterfaceSupported', args) ?? false;
+  }
+
+  @override
+  Future<List<BrowserProcessInfo>> getProcessInfos() async {
+    Map<String, dynamic> args = <String, dynamic>{};
+    final result = await channel?.invokeMethod<List<dynamic>>('getProcessInfos', args);
+    return result?.map(
+            (e) => BrowserProcessInfo.fromMap(e.cast<String, dynamic>()))
+        .whereType<BrowserProcessInfo>().toList() ?? [];
+  }
+
+  @override
+  Future<String?> getFailureReportFolderPath() async {
+    Map<String, dynamic> args = <String, dynamic>{};
+    return await channel?.invokeMethod<String>('getFailureReportFolderPath', args);
   }
 
   @override