+++ flutter_inappwebview_windows/windows/utils/base64.h
@@ -12,8 +12,8 @@
 #include <string_view>
 #endif  // __cplusplus >= 201703L
 
-std::string base64_encode     (std::string const& s, bool url = false);
-std::string base64_encode_pem (std::string const& s);
+std::string base64_encode(std::string const& s, bool url = false);
+std::string base64_encode_pem(std::string const& s);
 std::string base64_encode_mime(std::string const& s);
 
 std::string base64_decode(std::string const& s, bool remove_linebreaks = false);
@@ -25,8 +25,8 @@ std::string base64_encode(unsigned char const*, size_t len, bool url = false);
 // Requires C++17
 // Provided by Yannic Bonenberger (https://github.com/Yannic)
 //
-std::string base64_encode     (std::string_view s, bool url = false);
-std::string base64_encode_pem (std::string_view s);
+std::string base64_encode(std::string_view s, bool url = false);
+std::string base64_encode_pem(std::string_view s);
 std::string base64_encode_mime(std::string_view s);
 
 std::string base64_decode(std::string_view s, bool remove_linebreaks = false);